ZIN-96
- Rework python utilities scripts
- Rework completly menu.py
- Removed dependency to menuconsole that we used in the menu.py
- Added find_venv_activate_folder and find_venv_pip_path
- Removed  find_scripts_folder

+++ Tools/zinet_utilities/clear_build_folder.py
@@ -1,12 +1,12 @@
-import argparse
-import os
 import shutil
-import subprocess
-import sys
 
 from zinet_utilities.paths import find_zinet_root_path, find_zinet_build_path
 
 build_folder_path = find_zinet_build_path()
 print(f"Build folder path: {build_folder_path}")
 
-shutil.rmtree(build_folder_path)
+if build_folder_path.exists():
+    shutil.rmtree(build_folder_path)
+    print("Build folder cleared")
+else:
+    raise Exception("Build folder doesn't exist")
